update crate, add proxy test

+++ migrations/2017-12-25-120844_repo_base/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE repo_base (
 	id int8 PRIMARY KEY,
-	license_id int8,
+	license_id int4,
 	owner_id int8,
 	name text,
 	full_name text,
